backend/requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
pandas
python-dotenv
python-multipart
openpyxl
